+++ migrations/20250816200009_add_table_user_vibe.js
@@ -1,0 +1,17 @@
+const createLookupTable = require('../migration_utils/createLookupTable');
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.up = function (knex) {
+  return knex.schema.withSchema('users').createTable('user_vibe', createLookupTable);
+};
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.down = function (knex) {
+  return knex.schema.withSchema('users').dropTable('user_vibe');
+};