+++ src/util/getParseResultErrorMessage.ts
@@ -1,0 +1,2 @@
+export const getParseResultErrorMessage = <E>(parseResult: TODO) =>
+  parseResult.error.issues.at(0)?.message as E;